104 fix remaining failing tests (#106)
* fixed: chunking error
fixed: moved custom loading function to IO loading
fixed: updated syntax for tests

* small renaming changes

* fix: tests for denoising.py

* fix: tests for helper.py

* fix: tests for clustering.py

* small fixes

* remove use_opencv option

* remove border_nan argument for motion correct

* remove argument num_frames_split in motion correct

* - removing uint16 and 'half' datatypes to fix 'matching type' error

* - fix: default value for h5_loc in motion correction module

* - fix: No matching signature found

* - fix: No matching signature found

* - add: helper class to deal with parallel access of sample data

* - add: use new SampleInput class to retrieve sample data

* - add: SampleInput class can infer dataset location

* - fix: get h5_loc from SampleInput class

* - remove: removed conftest.py for better human readability

* - fix: enforce minimum length of 2 for DummyGenerator data

* - fix: appropriate error message if h5_loc not in .h5 file

* - fix: small changes to parameters in tests for IO loading

* - add: run all tensorflow tests in one worker

* - fix: change from relative error to absolute error for testing motion_correction_performance

* - fix: assertion statement of test_extension_full

* - fix: sequential testing for some functions

* - fix: split testing into parallel testing and sequential testing

* - update: updated sample data for better testing

+++ astroCAST/analysis.py
@@ -417,7 +417,7 @@ class Events(CachedClass):
         return events
 
     @wrapper_local_cache
-    def get_extended_events(self, video=None, dtype=np.half, extend=-1,
+    def get_extended_events(self, video=None, dtype=float, extend=-1,
                             return_array=False, in_place=False,
                             normalization_instructions=None, show_progress=True,
                             memmap_path=None, save_path=None, save_param={}):
